refactor(checkpoints): graduate to top-level layout, add docs-site page
Merge `main` and relocate from `experimental/checkpoints` to top-level
`pydantic_ai_harness/checkpoints/`, matching #347/#354. Drops the experimental
import warning, fixes import paths, moves tests. Adds the `docs/checkpoints.md`
page (nav.json, index.md, parity gate) and a README matrix row. No behavior change.
