+++ docs/nav.json
@@ -11,6 +11,7 @@
       { "label": "Compaction", "slug": "compaction" },
       { "label": "Overflowing Tool Output", "slug": "overflowing-tool-output" },
       { "label": "Step Persistence", "slug": "step-persistence" },
+      { "label": "Checkpoints", "slug": "checkpoints" },
       { "label": "Media", "slug": "media" },
       { "label": "Subagents", "slug": "subagents" },
       { "label": "Dynamic Workflow", "slug": "dynamic-workflow" },
