+++ pydantic_ai_harness/checkpoints/__init__.py
@@ -6,21 +6,18 @@ rewind/fork is out of scope for v1 -- that pairs with the branchable
 session-history track (harness issue #321).
 """
 
-from pydantic_ai_harness.experimental._warn import warn_experimental
-from pydantic_ai_harness.experimental.checkpoints._capability import (
+from pydantic_ai_harness.checkpoints._capability import (
     DEFAULT_BASH_TOOLS,
     DEFAULT_MUTATING_TOOLS,
     Checkpoints,
     CheckpointWarning,
 )
-from pydantic_ai_harness.experimental.checkpoints._shadow import (
+from pydantic_ai_harness.checkpoints._shadow import (
     Checkpoint,
     CheckpointError,
     CheckpointStore,
 )
-from pydantic_ai_harness.experimental.checkpoints._toolset import RestoreCheckpointToolset
-
-warn_experimental('checkpoints')
+from pydantic_ai_harness.checkpoints._toolset import RestoreCheckpointToolset
 
 __all__ = [
     'DEFAULT_BASH_TOOLS',
