+++ pydantic_ai_harness/checkpoints/_shadow.py
